fix: moves bitmap helpers to sdk-common package
