+++ modules/client/src/addresslistVoting/internal/client/estimation.ts
@@ -1,6 +1,5 @@
 import { AddresslistVoting__factory } from "@aragon/core-contracts-ethers";
 import {
-  boolArrayToBitmap,
   ClientCore,
   ContextPlugin,
   GasFeeEstimation,
@@ -10,7 +9,7 @@ import {
 } from "../../../client-common";
 import { IAddresslistVotingClientEstimation } from "../../interfaces";
 import { toUtf8Bytes } from "@ethersproject/strings";
-import { NoProviderError, NoSignerError } from "@aragon/sdk-common";
+import { boolArrayToBitmap, NoProviderError, NoSignerError } from "@aragon/sdk-common";
 
 /**
  * Estimation module the SDK Address List Client
